+++ index.html
@@ -35,6 +35,26 @@
         letter-spacing: 0.01em;
       }
 
+      .sr-only {
+        position: absolute;
+        width: 1px;
+        height: 1px;
+        padding: 0;
+        margin: -1px;
+        overflow: hidden;
+        clip: rect(0, 0, 0, 0);
+        border: 0;
+      }
+
+      .logo {
+        margin: 0;
+        font-family: "Courier New", Courier, "Liberation Mono", monospace;
+        font-size: clamp(0.55rem, 1.5vw, 0.8rem);
+        line-height: 1.05;
+        letter-spacing: 0.03em;
+        white-space: pre;
+      }
+
       header p {
         margin: 0 auto;
         font-size: 1.05rem;
@@ -179,73 +199,6 @@
         background: var(--maroon);
       }
 
-      .konami-overlay {
-        position: fixed;
-        inset: 0;
-        background: rgba(22, 15, 13, 0.86);
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        padding: 2rem 1.5rem;
-        z-index: 999;
-      }
-
-      .konami-dialog {
-        background: #ffffff;
-        padding: 1.5rem 1.5rem 1.75rem;
-        border-radius: 16px;
-        width: min(90vw, 560px);
-        box-shadow: 0 20px 45px rgba(0, 0, 0, 0.35);
-        text-align: center;
-      }
-
-      .konami-dialog h3 {
-        margin-top: 0;
-        margin-bottom: 1rem;
-        color: var(--maroon);
-        letter-spacing: 0.04em;
-        text-transform: uppercase;
-        font-size: 1.1rem;
-      }
-
-      .konami-frame {
-        position: relative;
-        width: 100%;
-        padding-top: 56.25%;
-        border-radius: 12px;
-        overflow: hidden;
-        box-shadow: 0 12px 35px rgba(80, 18, 20, 0.35);
-      }
-
-      .konami-frame iframe {
-        position: absolute;
-        top: 0;
-        left: 0;
-        width: 100%;
-        height: 100%;
-        border: 0;
-      }
-
-      .konami-close {
-        margin-top: 1.1rem;
-        padding: 0.55rem 1.25rem;
-        border-radius: 999px;
-        border: none;
-        background: var(--maroon);
-        color: #ffffff;
-        font-weight: 600;
-        cursor: pointer;
-        letter-spacing: 0.05em;
-        text-transform: uppercase;
-        transition: background 160ms ease, transform 160ms ease;
-      }
-
-      .konami-close:hover,
-      .konami-close:focus-visible {
-        background: #7d1c1f;
-        transform: translateY(-1px);
-      }
-
       @media (prefers-reduced-motion: reduce) {
         *,
         *::before,
@@ -260,8 +213,17 @@
   </head>
   <body>
     <header>
-      <h1>IEEE TXST Computer Society</h1>
-      <p>GitHub, GitHub Desktop, Tutorial Project.</p>
+      <h1 class="sr-only">IEEE TXST Computer Society</h1>
+      <pre class="logo" aria-hidden="true">
+███▀▀██▀▀██████▀   ▀██▀ ▄█▀▀▀█▄███▀▀██▀▀███   ▀████▀███▀▀▀███▀███▀▀▀███▀███▀▀▀███
+█▀   ██   ▀█ ███▄  ▄█  ▄██    ▀█▀   ██   ▀█     ██   ██    ▀█  ██    ▀█  ██    ▀█
+     ██       ▀██▄█▀   ▀███▄        ██          ██   ██   █    ██   █    ██   █  
+     ██         ███      ▀█████▄    ██          ██   ██████    ██████    ██████  
+     ██       ▄█▀▀██▄  ▄     ▀██    ██          ██   ██   █  ▄ ██   █  ▄ ██   █  ▄
+     ██      ▄█   ▀██▄ ██     ██    ██          ██   ██     ▄█ ██     ▄█ ██     ▄█
+   ▄████▄  ▄██▄▄  ▄▄███▄▀█████▀   ▄████▄      ▄████▄██████████████████████████████
+      </pre>
+      <p>Collaborate. Commit. Create.</p>
     </header>
     <main>
       <h2>Current Contributors</h2>
@@ -283,6 +245,6 @@
         >Learn more about GitHub Pages</a
       >
     </footer>
-    <script src="./js/script.js" defer></script>
+    <script type="module" src="./js/main.js"></script>
   </body>
 </html>
